this makes no sense

+++ 0x00-hello_world/101-quote.c
@@ -7,7 +7,7 @@
 int main(void)
 {
 	char b[] = "and that piece of art is useful\" - Dora Korpar, 2015-10-19\n";
-
-	write(2, b, sizeof(b)-1);
+	
+	write(2, b, (unsigned long)sizeof(b)-1);
 	return (1);
 }
